add analysis rule versioning
