+++ lunatrace/bsl/backend/src/hasura-api/graphql/get-manifest-dependency-edge-analysis-result.graphql
@@ -1,5 +1,5 @@
-query GetManifestDependencyEdgeAnalysisResult($vulnerability_id: uuid!, $manifest_dependency_edge_id: uuid!) {
-  analysis_manifest_dependency_edge_result(where: {vulnerability_id: {_eq: $vulnerability_id}, manifest_dependency_edge_id: {_eq: $manifest_dependency_edge_id}}) {
+query GetManifestDependencyEdgeAnalysisResult($vulnerability_id: uuid!, $manifest_dependency_edge_id: uuid!, $finding_source_version: Int!) {
+  analysis_manifest_dependency_edge_result(where: {vulnerability_id: {_eq: $vulnerability_id}, manifest_dependency_edge_id: {_eq: $manifest_dependency_edge_id}, finding_source_version: {_eq: $finding_source_version}, finding_type: {_neq: error}}) {
     id
     finding_type
   }
