+++ lunatrace/bsl/hasura/metadata/databases/lunatrace/tables/package_release.yaml
@@ -23,13 +23,6 @@ array_relationships:
         table:
           name: release_dependency
           schema: package
-  - name: release_dependents
-    using:
-      foreign_key_constraint_on:
-        column: dependency_release_id
-        table:
-          name: release_dependency
-          schema: package
   - name: release_licenses
     using:
       foreign_key_constraint_on:
