+++ lunatrace/bsl/ingest-worker/pkg/staticanalysis/rules/tpl/importedandcalled.yaml.tpl
@@ -5,23 +5,26 @@ rules:
     patterns:
       - pattern-either:
           - pattern-inside: |
-              $IMPORT = require("{{ .PackageName }}")
+              $IMPORT = require($PACKAGE)
               ...
           - pattern-inside: |
-              import $IMPORT from "{{ .PackageName }}"
+              import $IMPORT from "$PACKAGE"
               ...
           - pattern-inside: |
-              import * as $IMPORT from "{{ .PackageName }}"
+              import * as $IMPORT from "$PACKAGE"
               ...
           - pattern-inside: |
-              import { ..., $IMPORT,... } from "{{ .PackageName }}"
+              import { ..., $IMPORT,... } from "$PACKAGE"
               ...
           - pattern-inside: |
-              import { ..., $X as $IMPORT,... } from "{{ .PackageName }}"
+              import { ..., $X as $IMPORT,... } from "$PACKAGE"
               ...
+      - metavariable-comparison:
+          metavariable: $PACKAGE
+          comparison: $PACKAGE == "{{ .PackageName }}"
       - pattern-either:
-        - pattern-inside: $IMPORT.$FUNC()
-        - pattern-inside: $IMPORT()
+        - pattern-inside: $IMPORT.$FUNC(...)
+        - pattern-inside: $IMPORT(...)
     message: A vulnerable package was imported and called.
     languages:
       - javascript
