+++ lunatrace/bsl/semgrep/sandbox/imported-and-called.yaml
@@ -5,23 +5,26 @@ rules:
     patterns:
       - pattern-either:
           - pattern-inside: |
-              $IMPORT = require("vulnerable-package")
+              $IMPORT = require($PACKAGE)
               ...
           - pattern-inside: |
-              import $IMPORT from "vulnerable-package"
+              import $IMPORT from "$PACKAGE"
               ...
           - pattern-inside: |
-              import * as $IMPORT from "vulnerable-package"
+              import * as $IMPORT from "$PACKAGE"
               ...
           - pattern-inside: |
-              import { ..., $IMPORT,... } from "vulnerable-package"
+              import { ..., $IMPORT,... } from "$PACKAGE"
               ...
           - pattern-inside: |
-              import { ..., $X as $IMPORT,... } from "vulnerable-package"
+              import { ..., $X as $IMPORT,... } from "$PACKAGE"
               ...
+      - metavariable-comparison:
+          metavariable: $PACKAGE
+          comparison: $PACKAGE == "vulnerable-package"
       - pattern-either:
-        - pattern-inside: $IMPORT.$FUNC()
-        - pattern-inside: $IMPORT()
+        - pattern-inside: $IMPORT.$FUNC(...)
+        - pattern-inside: $IMPORT(...)
     message: A vulnerable package was imported and called.
     languages:
       - javascript
